Use ; to form composite instructions
This is so that control doesn't move ahead until task is done. For example asking user for PIN should stay on that line and have multi-turn conversation till user provides PIN. Having `YLD user` on a separate line means the control has already moved ahead when LLM gets called back and so doesn't go back to the ask even if user did not provide the info.